# daemon/DOCUMENTATION/openapi.yaml
openapi: 3.0.3
info:
  title: Minecraft Panel Daemon API
  description: API complète pour le daemon du panel de gestion Minecraft. Permet de gérer les utilisateurs, les serveurs, les conteneurs Docker, les fichiers, le système hôte, et l'historique/les sessions.
  version: 0.1.42
servers:
  - url: http://localhost:4000/api/v1
    description: Serveur Daemon local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
          nullable: true
...
    Automation:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        cron_expr:
          type: string
        action_type:
          type: string
        target_server:
          type: string
        payload:
          type: string
        created_at:
          type: integer
...
security:
  - bearerAuth: []
paths:
...
  /automations:
    get:
      tags:
        - Automatisations
      summary: Liste les tâches planifiées (Automations)
      responses:
        '200':
          description: Liste des automatisations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Automation'
    post:
      tags:
        - Automatisations
      summary: Créer ou mettre à jour une automatisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Automation'
      responses:
        '200':
          description: Automation créée
  /automations/{id}:
    delete:
      tags:
        - Automatisations
      summary: Supprimer une automatisation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Automation supprimée

  # SERVERS
...